gpt/requirements.txt
streamlit
flask
cx_Oracle
python-dotenv
google-generativeai